Amidst Us Class and Game Files
These are the files that have the actual game. Make sure to download the assignment instructions to see the directions I followed to create the game. The abstract Player.java parent class was given by the assignment but I wrote from scratch the Impostor and Crewmate interfaces; RedAstronaut and BlueAstronaut child classes, and the Gameplay driver file where the code has tests ready for you to run with one click.